requirements.txt
strawberry-graphql==0.258.0
Flask==3.0.0
sqlalchemy==2.0.23